requirement.txt
Django>=3.2,<4.0
djangorestframework>=3.12.4
djangorestframework-simplejwt>=4.8.0
pytz>=2021.1
sqlparse>=0.4.1
